Introduce GameContextType to properly handle the GameContext return type
