+++ src/components/Header/index.tsx
@@ -1,8 +1,7 @@
 import headerStyles from './index.module.css'
 
 import React from 'react'
-import { useGameContext } from '../../providers/game'
-import type { StatusProp } from '../../hooks/useGame'
+import { type StatusProp, useGameContext } from '../../providers/game'
 
 export default function Header (): JSX.Element {
   const [{ status, total }, { startRandomGame }]: [{ status: StatusProp, total: number }, { startRandomGame: () => void }] = useGameContext()
