+++ src/hooks/useGame.test.ts
@@ -8,13 +8,13 @@ describe('useGame hook', () => {
   it('has default data and mutations are properly handled', () => {
     expect(result.current[0].currentColumnIndex).toEqual(0)
     expect(result.current[0].currentRowIndex).toEqual(0)
-    expect(result.current[0].disabledKeys?.length).toEqual(0)
-    expect(result.current[0].exactMatches?.length).toEqual(0)
-    expect(result.current[0].grid?.length).toEqual(6)
-    expect(result.current[0].looseMatches?.length).toEqual(0)
+    expect(result.current[0].disabledKeys.length).toEqual(0)
+    expect(result.current[0].exactMatches.length).toEqual(0)
+    expect(result.current[0].grid.length).toEqual(6)
+    expect(result.current[0].looseMatches.length).toEqual(0)
     expect(result.current[0].solution).toEqual(['7', '*', '8', '-', '1', '2'])
-    expect(result.current[0].status?.complete).toEqual(false)
-    expect(result.current[0].status?.success).toEqual(false)
+    expect(result.current[0].status.complete).toEqual(false)
+    expect(result.current[0].status.success).toEqual(false)
     expect(result.current[0].total).toEqual(44)
     expect(result.current[1].deleteLatestEntry).toBeDefined()
     expect(result.current[1].handleValueOrOperatorClick).toBeDefined()
